make save

+++ libft/srcs/ft_printf/wchar_utils.c
@@ -50,7 +50,7 @@ char	*ft_wstr(int *arr)
 	while (arr[++counter])
 	{
 		symb = ft_wchar(arr[counter]);
-		str = ft_realloc(str, len, len + ft_strlen(symb));
+		str = ft_realloc(str, (size_t)len, (size_t)len + ft_strlen(symb));
 		len += ft_strlen(symb);
 		str = ft_strcat(str, symb);
 		ft_strdel(&symb);
